requirements.txt
opencv-python==4.5.1.48
tqdm==4.65.0
numpy==1.26.4
Pillow==10.0.1
texttable==1.6.4
PyYAML==6.0.1
git+https://github.com/openai/CLIP.git
torchvision==0.14.1
diffusers==0.30.2
transformers
gradio
pytorch==1.13.1
torchvision==0.14.1
pytorch-cuda==11.7
open_clip-torch
accelerate